Allow shift signup modal to be scrolled
Previously, if the list of shifts that are available extend beyond the current window's viewport, scrolling would only scroll the content behind the modal, not the modal's content itself.

To fix this, I'm adding overflow: scroll onto the overlay's styles. This is working for me on Chrome Desktop. I haven't tested it elsewhere, and I'm not convinced this is necessarily the right way to fix this.