Remove use of singleton exports to allow for more than one instance
While the use of `share` has already been removed, some files were
exporting singletons, such as the 'collections' object which meant that
initialising a second VolunteersClass instance overrode the collections.
This makes all instances of this into members of the class, and passes
the class into init functions for each.